Make CustomDropdown.xml offline capable
The widget is offline capable.
If I change in the 2nd line of CustomDropdown.xml offlineCapable="true" then I can use it in an offline app and it works.
If I don't add this Mendix Studio Pro complains:
Custom widget 'Custom Dropdown' is not offline capable and cannot be used on pages that are accessible through an offline profile.

In an offline app the data needs to be of course first synced to the offline app and then Nanoflow needs to be used instead of Microflows. But I used, tested and confirmed that it works correctly.